feat(http): allow filtering --since and --until dates
This is still a bit flakey, and doesn't really parse dates that are entered manually very nicely
yet. This will be easier once tests are written. Works well enough for now though.

One of the difficulties is that github always serves up the most recent commits for a given range,
so a bit of extra specificity is needed